+++ FXImageView/FXImageView.m
@@ -1,7 +1,7 @@
 //
 //  FXImageView.m
 //
-//  Version 1.2.1
+//  Version 1.2.2
 //
 //  Created by Nick Lockwood on 31/10/2011.
 //  Copyright (c) 2011 Charcoal Design
@@ -237,11 +237,11 @@
 #pragma mark -
 #pragma mark Processing
 
-- (void)setProcessedImageOnMainThread:(NSArray *)images
+- (void)setProcessedImageOnMainThread:(NSArray *)array
 {
     //get images
-    NSString *cacheKey = [images objectAtIndex:1];
-    UIImage *processedImage = [images objectAtIndex:0];
+    NSString *cacheKey = [array objectAtIndex:1];
+    UIImage *processedImage = [array objectAtIndex:0];
     processedImage = ([processedImage isKindOfClass:[NSNull class]])? nil: processedImage;
     
     if (processedImage)
@@ -259,7 +259,9 @@
         objc_msgSend(self.layer, @selector(addAnimation:forKey:), animation, nil);
         
         //set processed image
+        [self willChangeValueForKey:@"processedImage"];
         _imageView.image = processedImage;
+        [self didChangeValueForKey:@"processedImage"];
     }
 }
 
@@ -267,7 +269,7 @@
 {
     //get properties
     NSString *cacheKey = [self cacheKey];
-    UIImage *image = self.image;
+    UIImage *image = _originalImage;
     NSURL *imageURL = _imageContentURL;
     CGSize size = self.bounds.size;
     CGFloat reflectionGap = _reflectionGap;
@@ -276,14 +278,22 @@
     UIColor *shadowColor = _shadowColor;
     CGSize shadowOffset = _shadowOffset;
     CGFloat shadowBlur = _shadowBlur;
+    CGFloat cornerRadius = _cornerRadius;
+    UIImage *(^customEffectsBlock)(UIImage *image) = [_customEffectsBlock copy];
     UIViewContentMode contentMode = self.contentMode;
+    
+#if !__has_feature(objc_arc)
+
+    [_customEffectsBlock autorelease];
+    
+#endif
     
     //check cache
     UIImage *processedImage = [self cachedProcessedImage];
     if (!processedImage)
     {
         //load image
-        if (_imageContentURL)
+        if (imageURL)
         {
             NSURLRequest *request = [NSURLRequest requestWithURL:imageURL cachePolicy:NSURLRequestReturnCacheDataElseLoad timeoutInterval:30.0];
             NSError *error = nil;
@@ -299,7 +309,7 @@
                 image = [UIImage imageWithData:data];
                 if ([[[imageURL path] stringByDeletingPathExtension] hasSuffix:@"@2x"])
                 {
-                    image = [UIImage imageWithCGImage:image.CGImage scale:2.0f orientation:UIImageOrientationUp];
+                    image = [UIImage imageWithCGImage:image.CGImage scale:2.0f orientation:image.imageOrientation];
                 }
             }
         }
@@ -312,15 +322,15 @@
                                                        padToFit:NO];
             
             //apply custom processing
-            if (_customEffectsBlock)
+            if (customEffectsBlock)
             {
-                processedImage = _customEffectsBlock(processedImage);
+                processedImage = customEffectsBlock(processedImage);
             }
             
             //clip corners
-            if (_cornerRadius)
+            if (cornerRadius)
             {
-                processedImage = [processedImage imageWithCornerRadius:_cornerRadius];
+                processedImage = [processedImage imageWithCornerRadius:cornerRadius];
             }
             
             //apply shadow
@@ -334,7 +344,7 @@
             }
             
             //apply reflection
-            if (self.reflectionScale && self.reflectionAlpha)
+            if (reflectionScale && reflectionAlpha)
             {
                 processedImage = [processedImage imageWithReflectionWithScale:reflectionScale
                                                                           gap:reflectionGap
@@ -350,12 +360,17 @@
         {
             [[[self class] processedImageCache] setObject:processedImage forKey:cacheKey];
         }
+        [self willChangeValueForKey:@"processedImage"];
         _imageView.image = processedImage;
+        [self didChangeValueForKey:@"processedImage"];
     }
     else
     {
         [self performSelectorOnMainThread:@selector(setProcessedImageOnMainThread:)
-                               withObject:[NSArray arrayWithObjects:processedImage ?: [NSNull null], cacheKey, nil]
+                               withObject:[NSArray arrayWithObjects:
+                                           processedImage ?: [NSNull null],
+                                           cacheKey,
+                                           nil]
                             waitUntilDone:YES];
     }
 }
@@ -421,11 +436,17 @@
 
 - (void)updateProcessedImage
 {
-    UIImage *processedImage = [self cachedProcessedImage];
+    id processedImage = [self cachedProcessedImage];
+    if (!processedImage && !_originalImage && !_imageContentURL)
+    {
+        processedImage = [NSNull null];
+    }
     if (processedImage)
     {
         //use cached version
+        [self willChangeValueForKey:@"processedImage"];
         _imageView.image = ([processedImage isKindOfClass:[NSNull class]])? nil: processedImage;
+        [self didChangeValueForKey:@"processedImage"];
     }
     else if (_asynchronous)
     {
@@ -459,8 +480,11 @@
 
 - (void)setProcessedImage:(UIImage *)image
 {
+    self.imageContentURL = nil;
+    [self willChangeValueForKey:@"image"];
     self.originalImage = nil;
-    _imageView.image = ([image isKindOfClass:[NSNull class]])? nil: image;
+    [self didChangeValueForKey:@"image"];
+    _imageView.image = image;
 }
 
 - (UIImage *)image
@@ -470,12 +494,11 @@
 
 - (void)setImage:(UIImage *)image
 {
-    if (![_originalImage isEqual:image])
-    {
-        self.originalImage = image;
-        self.imageContentURL = nil;
-        
+    if (_imageContentURL || ![image isEqual:_originalImage])
+    {        
         //update processed image
+        self.imageContentURL = nil;
+        self.originalImage = image;
         [self updateProcessedImage];
     }
 }
@@ -601,11 +624,11 @@
 {
     if (![URL isEqual:_imageContentURL])
     {
-        //clear current image
-        self.imageContentURL = URL;
-        self.originalImage = nil;
-        
         //update processed image
+        [self willChangeValueForKey:@"image"];
+        self.originalImage = nil;
+        [self didChangeValueForKey:@"image"];
+        self.imageContentURL = URL;
         [self updateProcessedImage];
     }
 }