+++ FXImageView/UIImage+FX.h
@@ -1,7 +1,7 @@
 //
 //  UIImage+FX.h
 //
-//  Version 1.2.1
+//  Version 1.2.2
 //
 //  Created by Nick Lockwood on 31/10/2011.
 //  Copyright (c) 2011 Charcoal Design